pyproject.toml
[project]
name = "email-analyst-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "google-genai>=1.39.1",
    "logging>=0.4.9.6",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
